Journalio first commit :
This project is been developed together with the coding ninjas to help
make sure, the truth can be told and told in a way that cannot be
stopped and those that choose to tell these truthful stories can be
rewarded in cash or kind. We hope that journalists and reporters would
make a great living and even better would be for these awesome writters
to have a far more massive reader base.

Journalio ~ Truth in (a rewarding) light.